generator を core / integration / experiments に整理
